Break out most of gumppic functionality to agumppic.
And GumpPicWithWidth now inherits from AGumpPic - thanks @jeffboulanger
