+++ dev/Ultima/UI/Controls/GumpPicWithWidth.cs
@@ -8,7 +8,7 @@ using UltimaXNA.Core.Graphics;
 
 namespace UltimaXNA.Ultima.UI.Controls
 {
-    class GumpPicWithWidth : GumpPic
+    class GumpPicWithWidth : AGumpPic
     {
         private float m_PercentWidthDrawn = 1.0f;
 
@@ -32,8 +32,9 @@ namespace UltimaXNA.Ultima.UI.Controls
         }
 
         public GumpPicWithWidth(AControl parent, int x, int y, int gumpID, int hue, float percentWidth)
-            : base(parent, x, y, gumpID, hue)
+            : base(parent)
         {
+            buildGumpling(x, y, gumpID, hue);
             PercentWidthDrawn = percentWidth;
         }
 
@@ -42,7 +43,7 @@ namespace UltimaXNA.Ultima.UI.Controls
             Vector3 hueVector = Utility.GetHueVector(Hue);
             int width = (int)(m_PercentWidthDrawn * Width);
             spriteBatch.Draw2D(m_Texture, new Rectangle(position.X, position.Y, width, Height), new Rectangle(0, 0, width, Height), hueVector);
-            // base.Draw(spriteBatch, position);  - commented this out because we don't want to draw GumpPic's Draw method, but UGH we do want to call AControl's...
+            base.Draw(spriteBatch, position);
         }
     }
 }
