+++ dev/UltimaXNA.csproj
@@ -261,6 +261,7 @@
     <Compile Include="Ultima\Network\Client\GuildGumpRequestPacket.cs" />
     <Compile Include="Ultima\Network\Client\QuestGumpRequestPacket.cs" />
     <Compile Include="Ultima\Network\Server\ServerPingPacket.cs" />
+    <Compile Include="Ultima\UI\Controls\AGumpPic.cs" />
     <Compile Include="Ultima\UI\Controls\EquipmentSlot.cs" />
     <Compile Include="Ultima\UI\Controls\GumpPicWithWidth.cs" />
     <Compile Include="Ultima\UI\WorldGumps\ContextMenuGump.cs" />
